Base for dynamic dispatch established

+++ AbstractInstruction.h
@@ -12,7 +12,7 @@
 class AbstractInstruction {
 public:
     virtual ~AbstractInstruction() = default;
-    virtual void execute(SM83& cpu, MMU& mmu, uint8_t& cyclesDuringInstruction) = 0;
+    virtual void execute(SM83& cpu, MMU& mmu, int& cyclesDuringInstruction) = 0;
 };
